fix: area_info behavior
- area_info::ram_size is uint32 instead of size_t.
- get_next_area_info sorts the entries by the address instead of
the area_id.

+++ shared_headers/haiku_area.h
@@ -48,7 +48,7 @@ typedef struct haiku_area_info
     uint32 lock;
     uint32 protection;
     team_id team;
-    size_t ram_size;
+    uint32 ram_size;
     uint32 copy_count;
     uint32 in_count;
     uint32 out_count;
